Refactor modeling functions into separate file
